Move classes into packages to seperate concerns a little
