+++ jcl/src/main/java/io/mypojo/jcl/ClasspathResources.java
@@ -454,4 +454,26 @@ public class ClasspathResources {
     public void setIgnoreMissingResources(boolean ignoreMissingResources) {
         this.ignoreMissingResources = ignoreMissingResources;
     }
+
+    public static class JclJarEntry {
+
+        private String baseUrl;
+        private byte[] resourceBytes;
+
+        public String getBaseUrl() {
+            return baseUrl;
+        }
+
+        public void setBaseUrl(String argBaseUrl) {
+            baseUrl = argBaseUrl;
+        }
+
+        public byte[] getResourceBytes() {
+            return resourceBytes;
+        }
+
+        public void setResourceBytes(byte[] argResourceBytes) {
+            resourceBytes = argResourceBytes;
+        }
+    }
 }
